rename : Kiosk_Lv4 폴더로 이동
Kiosk_Lv3 -> Kiosk_Lv4
