+++ src/kiosk/Kiosk_Lv4/Kiosk.java
@@ -1,6 +1,5 @@
-package kiosk.Kiosk_Lv3;
+package kiosk.Kiosk_Lv4;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
