+++ src/kiosk/Kiosk_Lv4/MenuItem.java
@@ -1,4 +1,4 @@
-package kiosk.Kiosk_Lv3;
+package kiosk.Kiosk_Lv4;
 
 // 음식 항목 관리
 public class MenuItem {
